pyproject.toml
[project]
name = "sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Joon-Hyuk Ko", email = "jhko725@kias.re.kr" }
]
requires-python = ">=3.13"
dependencies = [
    "colorcet>=3.1.0",
    "diffrax>=0.7.0",
    "dynamical-systems",
    "dynamics-discovery",
    "dysts>=0.95",
    "equinox>=0.13.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=6.30.1",
    "jax[cuda12]>=0.7",
    "jaxtyping>=0.3.2",
    "lyapunov",
    "matplotlib>=3.10.5",
    "matplotlib-utils",
    "more-itertools>=10.7.0",
    "numpy>=2.3.2",
    "optax>=0.2.5",
    "orbax-checkpoint>0.5.0",
    "ott-jax>=0.5.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.1",
    "tqdm>=4.67.1",
    "wandb>=0.21.3",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
lyapunov = { workspace = true }
matplotlib-utils = { workspace = true }
dynamical-systems = { workspace = true }
dynamics-discovery = { workspace = true }

[tool.ruff]
extend-include = ["*.ipynb"]
src = []
fix = true

[tool.ruff.lint]
fixable = ["I001", "F401"]
ignore = ["E402", "E721", "E731", "E741", "F722"]
select = ["E", "F", "I001"]

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2
order-by-type = false
